allow missing questions
instead of throwing an error if a question or answer is missing (or if
there are more questions in a category than dollar values), log a
warning or error and render a blank cell for that question.

this commit adds function `getData` that handles logging if some part of
the data is missing.